Rename: CollectionButtonStrokedSvg.tsx →PickUpButtonStrokedSvg.tsx, CollectionButtonTintedSvg.tsx →PickUpButtonTintedSvg.tsx
