+++ src/components/atoms/svg/PickUpButtonStrokedSvg.tsx
