+++ src/components/atoms/svg/PickUpButtonTintedSvg.tsx
